move losses to shared
