+++ demo/lightning/image-segmentation/model.py
@@ -18,7 +18,7 @@ import torch
 import time
 import lightning.pytorch as pl
 from unet3d import Unet3D
-from losses import DiceCELoss
+from demo.shared.losses import DiceCELoss
 
 
 class Unet3DLightning(pl.LightningModule):
